Be explicit about whether we're parsing a systemlib unit or not
Summary:
unit-parser.cpp was checking !SystemLib::s_inited to determine if the
unit being parsed belongs to systemlib or not. If we've disabled
systemlib parsing, then this flag will never be set, thus everything
will be considered as part of systemlib. Instead of trying to infer it
this way, make it an explicit parameter of the parsing path.

Reviewed By: oulgen

Differential Revision: D34516408

fbshipit-source-id: bd0617f10c0e97401282b4b2e2f457b78cf1e90b