Don't call _Exit from within unit-parser.cpp
Summary:
With the right runtime options, unit-parser.cpp can emit an error then
call _Exit if we get an ICE unit. This is dubious to begin with, but
its especially bad if we're using subprocesses to do the parsing. The
subprocess will exit, but not the main parsing process, defeating the
purpose. Instead of calling _Exit, throw a special exception. The
caller can catch the exception and do whatever it wants. Right now the
two callers will catch it and call _Exit as before.

Reviewed By: oulgen

Differential Revision: D34516433

fbshipit-source-id: d5ff7db06613d411301716a3226a6bab4b8c17bb