hhbc-gen
Summary:
This is a demo of two tools we might want for HHBC ffi.

* gen-rust.cpp: a C++ program that macro-generates hhbc.rs containing a Rust `enum Op`
with a variant for each bytecode instruction. This translation is lossless - the enum contains
all the information present in the OPCODES macro. The Rust enum can be used by the Emitter
and cbindgen.
* gen-cpp.rs: a Rust program that translates `enum Op` losslessly back to a C++ OPCODES macro.
This just completes the round trip so we know both translation directions work without losing information.

gen-rust.cpp would let us keep the definition in C++ and could enable using Rust procedural macros to replace boilerplate in emitter/instruction_sequence.rs and bytecode_printer/print.rs.

Reviewed By: kavoor

Differential Revision: D34502363

fbshipit-source-id: 7fe593477e0de709311d52f0360eb069f3da8a45