Update SymbolMap::getAllTypes() and company so they don't have to wait for the DB to finish updating
Summary:
I implemented `SymbolMap::getAllTypes()` by waiting for the DB to finish updating, then querying the DB for all of its types. This will not work in a world where the DB is networked - a networked DB is pinned to a specific public commit, and will never update.

Instead, we need to query the DB for all its types, then override those types with the information we have in-memory, representing the local changes that have happened since the DB's public commit. That's what this diff accomplishes.

Reviewed By: ricklavoie

Differential Revision: D34570621

fbshipit-source-id: bc5728ffaa989af4072fc98c9d29595d4fa8865e