Lambda whose expected type is supportdyn of non-function type
Summary:
When an untyped lambda is passed to a `mixed` or `dynamic`, we check the body assuming that the untyped parameters are `mixed` or `dynamic` respectively.

Extend this to expected type `nonnull`; in this case, untyped parameters are assumed to be `mixed`.

Also extend to `supportdyn<mixed>` and `supportdyn<nonnull>`; in this case, untyped parameters are assumed to be `supportdyn<mixed>`. (Otherwise, we would lose errors when pessimising e.g. trying to invoke a method on a mixed.)

Reviewed By: vassilmladenov

Differential Revision: D34552174

fbshipit-source-id: 75cf80b7e6791b0b85e097a742f58e56435f0d51