If a hh_compile fails to compile a file it should return a non-zero status
Summary:
Previously if a file failed to compile the hh_compile command would return status 0 as long as the file could be read properly.  This changes it so if one or more of the files passed in fail to compile then it continues to process the rest but when compilation is done it returns an error.

Also did some drive-by `anyhow::Result` -> `Result` cleanup.

Reviewed By: edwinsmith

Differential Revision: D34535341

fbshipit-source-id: cf8d71dba94e2e6e0855b97cb64c0378e1c5f632