Add support for tracking flows to/from properties on objects
Summary:
As title. This is a pretty basic model that should work fine for most
cases. We keep a map of tainted properties per object. From what I can tell,
the main mistaints here will be:

1) Objects with custom __get or __set implementations (no idea how common this is)
2) There are certain cases where setM can throw, we don't unwind properly

We can address those as they come up.

I also left a TODO for optimization: for some reason doing the naive fix
doesn't work - can debug that in the future if we run into high memory usage

Reviewed By: dkgi

Differential Revision: D34542983

fbshipit-source-id: 64f50ea4e6e1b27cc9110a58ab628de92062c50b