Add asserts to AwaitAllWaitHandle destructor
Summary:
Debugging one of the AsioBlockableChain::exitContext() segfault core, it is
apparent from refcounts that an AwaitAllWaitHandle object was freed, while one
of its children was not yet finished. At the same time, the AAWH was not
removed from that child's parent chain.

Let's add an assert that will hopefully fire and point to the erroneous decref.

Reviewed By: alexeyt, mofarrell

Differential Revision: D34590208

fbshipit-source-id: 95467d00bb572a9c91261b1087a09c5b0d92ece3