Change hhbc.rs to be a more structured data format.
Summary:
Previous gen-hhbc was producing something like kind of looked like hhbc_ast's Instruct enum - but wasn't actually usable in that form - so we ended up with all the restrictions of a Rust enum without actually any benefit.

This changes gen-hhbc to emit a structured table (P483915232) that we can use as input to various macros to generate code from (see the next diff).

In the future if we switch to the Rust file as the source-of-truth for the opcodes we can switch to a annotated enum that actually defines the Instruct enum and then generates the data table as a derive macro from that.

Reviewed By: edwinsmith

Differential Revision: D34584033

fbshipit-source-id: c3aa74b27dd9072dcd8d7d8d54f1e1bbb7a49546