Fix JIT time type pessimization during hoistCheckTypes
Summary:
Change hoistCheckTypes to hoist the CheckType through other operations in a
block, rather than hoist all operations in the block into predecessors.

In general it is not safe to hoist AssertType into predecessors.  This can
worsen types violating IR op input requirements.  This is because the jit
type union is imprecise and pessimistic.  The AssertType may simplifiy away
in predcessors, while the join block may end up with a type no longer
satisfying the AssertType because of the pessimistic type union.  This same
issue does not apply to CheckType, as CheckType only checks for types that are
precisely representable by the jit type system.

It is worth noting that sinking AssertTypes is fine.  They can be moved into
the next and taken blocks of a subsequent CheckType (which is the end effect of
this change).

Reviewed By: ricklavoie

Differential Revision: D34590042

fbshipit-source-id: d73f5fbb2c05fc54a459e8c4e21a0128ede62e10