Fix refcounting bug during onAwaitAllCreate profiling event
Summary:
Normally, a returned AwaitAllWaitHandle has a refcount of two. One for the
return value, the other for AAWH not being finished yet, which accounts for
all edges from its children.

If a profiler is set up to handle the onAwaitAllCreate event and the profiler
finishes the AAWH (e.g. by calling HH\Asio\join()), then it will cause the
refcount to drop by one.

Previously, we did the second incref after the profiling hook, so the profiler
described above would free the AAWH and the AAWH construction would then return
a garbage.

Fix this by increfing prior to the profiling hook.

Reviewed By: ricklavoie, alexeyt

Differential Revision: D34630132

fbshipit-source-id: abf0b29edf004a2435411fb4cf4505569163c036