Remove separate CollectionsHeap
Summary:
The locals heap is already used for tracking values. Reuse that for
now. This reduces complexity, but also helps with correctness since the actual
interpreter does not make any distinction here. This came up when trying to
implement iterators (see next diff).

We'll face some pain in the future when we try to be more precise for
collections, but this is fine for now.

Reviewed By: dkgi

Differential Revision: D34599316

fbshipit-source-id: 73ccb2aa5729480fa726ae447f208ba04e33e94a