Unify `unset` lval checks with other lvals
Summary:
**Context:** I'm trying to change Hack to treat lvals more consistently, with the ultimate goal of encoding lvals distinctly within Hack itself.

The core change here is that we now use the same set of checks for the arguments of `unset` as we do for other lvalue like bits of syntax.

In practice, this diff *mostly* is just error message re-arranging. As it stands, most of these cases were already caught at HackC time (see: HHVM tests). The only new case I see is that we disallow `unset(varray[][0])`.

I've started threading through the "root lval" for this check: this lets us give better error messages for code like:
```
$this++;
```

Reviewed By: Wilfred, edwinsmith

Differential Revision: D34189292

fbshipit-source-id: f358c3a31d524fe166050e107d6e84948c835918