define LocalRange
Summary:
Several HHBC instructions take a LocalRange (aka LAR) operand that
encodes a contiguous sequence of LocalIds as a (start, len) pair.
Since this utility type exists in HHVM, lets define it here too.
In addition, in HHAS and HHBC, empty ranges are encoded as `L0:0`,
which is not really a sentinel value, so we can save a tag word and
simplify hackc by not using Maybe.

LocalRanges can only be unnamed locals, so use LocalId instead of
Local and propagate that invariant through the emitter and assert
early if we get it wrong somehow.

Reviewed By: aorenste

Differential Revision: D34200673

fbshipit-source-id: e740fe42fe6fa33b58ac5128f10a0771ecd65efd