Initial support for builtins
Summary:
Support builtins as tito.

We need to smuggle some state from the `FCall*` opcodes so we can track
arguments properly, which I achieve via the state object. Once that's set up,
we can do most of the things we want just like we do for normal functions.

This does not yet handle inout parameters (I need to read more code for that) -
in an ideal world we'd taint the inout params via tito as well.

This also does not yet support constructs like `Vec\map` (we need to handle
closures) or things like `(Set {})->firstKey()` (because the `this` pointer
needs to be propagated properly). But it's a start!

Reviewed By: dkgi

Differential Revision: D34660803

fbshipit-source-id: 56da2cfe781548464fff8569abca3c2624fa1026