Modify some types to be inline with hhbc.h
Summary:
- `Switch` should take i64 as base
- Many args take an IVA aka u32 instead of usize.
- Reordered FuncCallArgs constructor based on order of fields in struct defintion and order they are printed.

`StackIndex` and `NumParams` are mixed around a bit in `emit_expression.rs`. Some of the `stack_size` values we plug into `BaseSC` we also plug into `QueryM`.
`QueryM` is specified to take `NumParams` and `BaseSC` is specified to take `StackIndex`.

I think `StackIndex` makes the most sense for all of these as in the `QueryM` case it means "number of values to pop before we push result". ([RFC}?)

Reviewed By: aorenste

Differential Revision: D34648981

fbshipit-source-id: f2d4945f8656130aac12a982f695144d049eb43e