improve error message and position for "bad decl override" error
Summary:
See the fixed test cases below for examples of the change.
Now that we have wired down AST positions of a class' immediate parents, we can have a more precise position for some of the errors produced by Typing_extends.
In this diff, the primary position for "bad decl override" error is now on the implemented interface instead of the class implementing it.

Differential Revision: D34583307

fbshipit-source-id: 24aff361e542cb2f70f2a8d8dd041a58f52dee39