Switch to global Consers
Summary:
Rather than requiring callers to invoke `Conser::mk` on some `Conser` instance in order to hashcons a value, add a `Consable` trait defining a static storage location for the single global `Conser` table for any type which can be hashconsed. Use `Consable` to add a simpler interface for constructing an `Hc`, `Hc::new`. Since this provides a way for impls on `Hc<T>` to access the appropriate `Conser`, it allows us to implement deserialization without needing to set the `Conser` to a thread-local (or similar).

I removed the support for unsized types like `str` (rather than designing an interface for it) because it is currently unused: we use the `intern` crate for strings and byte strings instead.

Reviewed By: edwinsmith

Differential Revision: D34659239

fbshipit-source-id: 2f0cdcc984027cb2af0144c1b6290749cc60ff0d