invert the sense of happly_decl_has_no_reified_generics()
Summary:
This is the first place in hackc we call a DeclProvider to resolve external
classnames. I started to refine it's api to be a tri-state so we can distinguish
"I don't know" cases from "definitely yes" and "definitely no", but managed
to hopelessly confuse myself with double negatives.

This diff flips the sense of the API to match the comment, and adds some
asserts and exhaustive pattern matches, but should result in no bytecode changes
(with or without DeclProviders).

Reviewed By: aorenste

Differential Revision: D34617436

fbshipit-source-id: a0f581a243edeb95b432f6c87f5291014e3ec421