Use folded decls for the default typecheck tests
Summary:
Previously we would use shallow decls for the `typecheck:typecheck` tests, and the `typecheck:typecheck_legacy_decl` reflected how type checks are done in practice.

Make `typecheck:typecheck` match how hh runs, and introduce a `typecheck_shallow_decl` for a test that exercises the shallow decl code paths.

This is not a logic change, it's only clarifying which tests exercise which mode.

Differential Revision: D34434739

fbshipit-source-id: 0bf20857f6909f0ad0856ea120f8c3766755d753