Change debug printing of folded decls so that hashmaps are sorted
Summary:
I want to use the printed decls as sort of a verify.py unit test for fold.rs changes. But currently, since the iteration order of HashMap and HashSet are arbitrary, the output of the debug print changes on each run, which doesn't make for a very good unit test. Instead, I want to just sort the fields by key value and then print them.

I attempted making this a little trait method on the maps, but realized that since I don't own those types, it's not easy to do so, so created some helper methods to do that. `sorted_set` will be on the next diff.

The output of the print is modified a little bit, but I'm not concerned unless someone is invested is exactly how the output is emitted.

Differential Revision: D34703470

fbshipit-source-id: 8e043bcac044f50f6e41671c7baf5414fd8d50a4