req_ancestors_extends and req_ancestors
Summary:
This is the OCaml code I'm using as reference:

The two fields: `dc_req_ancestors` and `dc_req_ancestors_extends`

https://www.internalfb.com/code/fbsource/[74eb9fd56b64]/fbcode/hphp/hack/src/decl/decl_defs.ml?lines=184-185

The `requirement` type:

https://www.internalfb.com/code/fbsource/[74eb9fd56b64]/fbcode/hphp/hack/src/typing/typing_defs.ml?lines=128-142

The call to populate those fields:

https://www.internalfb.com/code/fbsource/[74eb9fd56b64]/fbcode/hphp/hack/src/decl/decl_folded_class.ml?lines=936-937

The logic that is being triggered:

https://www.internalfb.com/code/fbsource/[74eb9fd56b64]/fbcode/hphp/hack/src/decl/decl_requirements.ml?lines=132-164

I've put all of the logic in `decl_requirements.ml` into `fold.rs` rather than breaking it out as a separate file as it is the only caller of `decl_requirements` that I can find. Let me know if separating the logic out would be more desirable.

Reviewed By: shayne-fletcher

Differential Revision: D34589318

fbshipit-source-id: 2eeba87929d897c10f3ed97a8cce02b758dd1073