Add a CI verify.py test for folded decls
Summary:
We should start testing folded decl logic. This adds a CI test based on verify.py to exercise the folded_decl logic and print out results courtesy of `decl_file`.

I'm temporarily moving the `tast` tests into their own directory. Even though I believe those are supposed to compare decls as well, some of the features are not implemented yet. When we get to the point where we are nearing feature complete we can either combine the two directories and tests or get rid of a test.

Reviewed By: edwinsmith

Differential Revision: D34626181

fbshipit-source-id: e35e1c6a4bfb4b56fe951d876e9304fd3a7c1442