Send a minimal text edit over LSP
Summary:
VS Code does not try to preserve the cursor position when you modify a large amount of text. We seem to exceed this limit on files that are >3000 LOC or >96 KiB. This is frustrating for users: their cursor jumps to the beginning/end of a file when they save.

Don't send unchanged lines in the LSP text edit. This means that we send much smaller responses and VS Code still preserves cursor position.

Reviewed By: ljw1004

Differential Revision: D34590469

fbshipit-source-id: 2a4d40b4644a865779d02ff23efeb14e35ec564e