Don't use wordcpy in hugifyText
Summary: Clang can turn a call of wordcpy into a call to memcpy, which is likely to be one of the hot functions.  Instead, use a special new function to do the copy, and put annotations to prevent it from being turned into memcpy and to be placed in the middle of the hot functions.

Reviewed By: oulgen

Differential Revision: D34722554

fbshipit-source-id: 09650ed9c65bd6391cdea4897b6fff5dd79fe988