Fix bug in setting variances on call parameters with like types
Summary:
Consider the following.
```
function my_reduce<Tv as dynamic, Ta as dynamic>(
  ~Traversable<Tv> $traversable,
  ~(function(Ta, Tv): Ta) $accumulator,
  ~Ta $initial,
) : void {}

function f(~dict<string, ?bool> $d) : void {
  my_reduce($d, ($x, $y) ==> $x ?? $y, null);
}
```
To infer types for the call to `my_reduce`, the type variables representing `Tv` and `Ta` need to get their variances set according to how they are used in the type of the `$accumulator` parameter. This diff makes that code look through the `~` on the type to set them properly. Previously, the variances weren't getting set, causing inference to fail later by solving an incorrectly-not-marked-as-contravariant type variable used in checking `??` prematurely solve to nothing.

Reviewed By: andrewjkennedy

Differential Revision: D34752053

fbshipit-source-id: da769c466efffedcd55e8678beec807b45cd4b03