Remove --disallow-trait-reuse flag and functionality
Summary:
In light of the recent work around diamond trait inclusion - it does not seem like we will ever be in a position to turn this feature on. I want to remove this to both clean up the codebase a bit and also so that we carry fewer things over into the rust implementation.

This flag was originally introduces in 2020 in D21441638 (https://github.com/facebook/hhvm/commit/c54aa1a09f474cc7dd62af6923bca0fcac8eae42) and covers a feature that has been aspirational since 2016.

Reviewed By: hgoldstein

Differential Revision: D34694803

fbshipit-source-id: 8be3041af7ed02f459a058a31f52834e46c11f91