index namespaces locations
Summary:
We want to index Namespaces declarations locations in Hack code. That is, generate glean facts containing the name and source code location of the namespace declarations.

The `tast` that the indexer is using to generate facts doesn't contain `Namespace` nodes. We get them from the cst that we retrieve from the provider_context entry.

Note that there are several asts available from the provider_context entry.
1. the CST
2. the Full fidelity AST

Ideally, it would have been slightly simpler to get the ns declarations from 2 as the AST is easier to parse, but because it was built with `elaborate_namespaces:true`, the namespaces declaration nodes have been replaced with `SetNamespaceEnv` nodes which don't contain span information.

I've considered various solutions:
1. add span information in SetNamespaceEnv
2. Use `elaborate_namespaces:true` by default when constructing the ff_ast
3. re-compute a full fidelity ast with `elaborate_namespaces:false`
4. use regex
5. use the CST

1 and 2 require changing the hack ast API. 3 was inefficient. 4 less robust. I settled with 5.

Reviewed By: nhawkes

Differential Revision: D34406294

fbshipit-source-id: 7e2c74a7aa71499f08aeaed63fa3dea8208bef0c