Stop dune from depending on hackc
Summary:
HackC is about to start depending on a generated file in tools/hhbc-gen.  We can generate the file in buck and cmake.

I don't think anything in dune should depend on HackC anymore - and this passes tests so it's better to remove the dependency rather than trying to convince dune to build the file.

Reviewed By: kavoor

Differential Revision: D34776630

fbshipit-source-id: 4cefee316413065b66d5c6eabcdf6ec92f8c7277