Expose the ability to convert an object into an array while ignoring late init
Summary:
The current object_prop_array is terrible. No question about it. Unfortunately, it's not terrible enough. It doesn't currently handle the case of converting an object to an array when that object may contain late init properties that aren't initialized.

This exposes that functionality for the few cases that need it.

Reviewed By: ricklavoie

Differential Revision: D34826602

fbshipit-source-id: 9b7a5316c53bd4d7e9c3abd1f9dec7a083a70a32