Fix memory accounting when not using JEmalloc
Summary:
When not using JEmalloc (for example if we're compiling under ASAN),
we don't track memory usage, thus none of our OOM checks really work
(usage is always 0).

Instead of only using allocation counters when JEmalloc is enabled,
always use them. When using JEmalloc, the counters work as they did
before. When not using JEmalloc, we manually adjust the counters when
allocating and freeing.

Reviewed By: edwinsmith

Differential Revision: D34837790

fbshipit-source-id: 0d5ac319b9ac19ca70be2656abd94da1c6457f83