Resolve tyvars
Summary: Adds the equivalent of `expand_tyvar` to the inference env; this allows us to follow the bindings of a type variable until we reach a concrete type, a dead end or an error state i.e. a circular ref between tyvars. In the non-error cases, we also shorten the paths of aliasing tyvars.

Differential Revision: D34788852

fbshipit-source-id: 476cb845c8a9ed65047b779ed498d9582ca84444