Fix bug in OCaml naive_dedup and Rust flatten_parent_class_reqs
Summary:
Make our requirements test case a bit more complicated, so as to expose (and fix) a bug in the OCaml and a bug in the Rust.

In OCaml, we are failing to dedup at all: we accidentally included a position in our hash table by making the keys `pos * string` rather than just `string`. Switch to a `String.Table` (which can raise `Not_found_s` in addition to `Caml.Not_found`) to allow this function to actually dedup requirements.

In Rust, we're reversing an extra time in flatten_parent_class_reqs.

Reviewed By: shayne-fletcher

Differential Revision: D34877262

fbshipit-source-id: d29b8c804fae24fd23f673700243badbb83cdd1f