Introduce extern-worker framework (#9025)
Summary:
Pull Request resolved: https://github.com/facebook/hhvm/pull/9025

Introduce the new extern-worker framework. This is a framework which
allows you run work out of process. You "upload" data into the
framework and receive an abstract "Ref" to the data. You can then
invoke various "jobs" with those Refs as inputs,and receive more Refs
as outputs. This lets you pipe data from one job to the next without
needing to have it be present locally.

It supports different "implementations" which can store the data and
execute the jobs different ways. The implementation which is always
available stores data on disk and uses fork+exec to execute the
jobs. Most of the framework is agnostic to the implementation, keeping
the logic an implementation needs to implement to a minimum.

For extra reliability, the framework supports "fallbacks". Depending
on configuration, if executing a job using the normal implementation
fails, we can retry the job using the "fallback" implementation, which
is always the fork+exec implementation. This (might be) slower, but is
considered more reliable, and is designed to prevent spurious failures
from stopping execution. This fallback is invisible to the user of the
framework.

For the most part, jobs can be written as normal functions within a
C++ class. They receive/return normal C++ types and don't have to
worry about serialization or how the data is transported around. The
only requirement is that the types must be usable with
BlobEncoder/BlobDecoder.

Implementations besides the builtin fork+exec one can be provided by a
creation hook. This hook can be overwritten and used to register a
creation function. So, additional implementations can be separate from
the framework proper.

Reviewed By: mofarrell

Differential Revision: D34710025

fbshipit-source-id: 7b180bad60210c4814a33f813772ccb2c0d01b31