Allow type projection on tyvar in dynamic mode
Summary: Currently we reject type projections on type variables that have no lower bounds, at the variable solving stage. Unfortunately this results in a lot of errors in dynamic mode checking (see test), because often we can infer nothing useful about a type variable that participates in a type projection. It's not clear to me why we need to reject type variables so late - aside from type constants, we can solve type variables however we wish, in accordance with bounds - but during dynamic checking mode it seems benign. We should revisit this.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D34898380

fbshipit-source-id: c3f62417d286f57071d2b23bfa3ddd198069a4a0