pessimised builtins were not compatible with like pushing
Summary:
Previously, this had a type error running under pessimised builtins.
```
function f () : shape('a' => vec<int>) {
  return shape('a' => vec[4]);
}
```
The problem here is that like-pushing and the way vec[] expressions work (under pessimised builtins) with expected types are not compatible. We pessimised the expected type of a to `~vec<int>`, but then we like-push the expected type so that the expected types of the elements of the `vec` are `~int`, and then we use that as the resulting type instead of `int`, so we get `shape('a' =>vec<~int>)`. If we didn't like-push the `vec`, we'd expect `int` for it's elements, and get that (and if there were some `~int` arguments, pessimised builtins would take care of that on its own).

So we just make sure not to use like-pushing for pessimised builtins.

Reviewed By: andrewjkennedy

Differential Revision: D34897848

fbshipit-source-id: ba430a85edf933e9481e6cb6449b9deeefbae3c5