Return supportdyn field type from Shapes::idx on supportdyn shape
Summary:
When applying `Shapes::idx` to a value of type of the form `supportdyn<shape(_)>`, we should preserve `supportdyn` in the result. Not only for missing fields (so we get `supportdyn<mixed>` instead of `mixed`) but also listed fields that don't support dynamic.

To do: simplify result type. Because of the way that we construct the result using a type variable, simplification doesn't happen e.g. we might see `supportdyn<int>`. This is benign but unergonomic.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D34923177

fbshipit-source-id: 26761177a3105388ab1d9c66711199f503c5aeab