Fix foreach over dynamic typed arrays
Summary: Previously, we detected whether the array was type `dynamic` be checking if it was a subtype of `dynamic`, but that was too permissive when dynamic-aware subtyping was being used. For example, `vec<int>` is a subtype of `dynamic` in dynamic-aware mode, but should not be treated by `foreach` as being `dynamic`. This adds a check that dynamic is a subtype of the type. We only check in the case that dynamic-aware subtyping is being used.

Reviewed By: andrewjkennedy

Differential Revision: D34931956

fbshipit-source-id: 539d8cbdf0fd866f54b847797d280f73d1646342