save stack usage by factoring out more match arms
Summary:
This helps somewhat (1M is nothing to sneeze at), but the main stack usage
in parse_term() is the numerous compiler-generated copies of `parser1`,
due to the style of using a temporary cloned parser to look ahead one token.
Each transient copy of `parser1` is 192 bytes.

This coding style is pervasive, not much we can do about it without more churn
or using a different technique for lookahead.

Reviewed By: shayne-fletcher

Differential Revision: D34926523

fbshipit-source-id: 9caa8ae233f7cd152f3c1b81ae72767abda6b1e3