delay-init 3/4 - add initializationOptions.delayUntilDoneInit
Summary:
This diff adds a new initializationOption:
```
    delayUntilDoneInit: bool;
        (** used for test scenarios where we want clientLsp to delay receiving any
        further LSP requests until clientIdeDaemon has done its init. *)
```

The way I've implemented it is, if the flag is set, then clientLsp will queue up everything that arrives from the LSP client until after serverless-IDE is either ready or stopped (rather than the existing behavior, which is to respond to it immediately with "sorry cannot do")...
* It first claims to be ready upon initialization, once it's finished indexing the last file it needs, and its sends `Done_init {total=0}`.
* Subsequently it will become non-ready if it receives didChangeWatchedFiles, and then become ready again once it has finished processing the changes.

The reason for this? It just makes it easier to write tests. We can send the initialize message and then immediately send other messages, without having to write code to wait until it's ready.

Reviewed By: Wilfred

Differential Revision: D34849432

fbshipit-source-id: 1951eacb6dc2889df2fe27d0d43fd6f31ca6892d