fix jsonrpc race
Summary:
The behavior of jsonrpc.ml is that
1. There is a daemon which reads json from stdin, stores it in a queue, and sends contents of that queue over a pipe whenever (1) it doesn't have anything to read on stdin, (2) the pipe is available for writing. The reason is so that it's always available to read from stdin should there be anything there, so that it can get as accurate as possible timestamps about when the message was sent by VSCode.
2. The main process, whenever it's curious about the next message, will read everything it can from the pipe into a queue and then handle (in clientLsp) the first message in the queue. The reason for having this queue is in case there were any messages relating to cancellation, so it can act upon cancellation promptly. For instance when we pick up message ID 5, we want to scan through the entire queue to see if there's a cancellation ahead of us for ID 5. This requires us to have the queue in a data-structure, rather than in an OS pipe buffer.

This diff concerns a race condition where the daemon encountered a series of messages over the pipe ending with an EOF, for instance `hover` then `shutdown-request` then `exit-notification` then EOF. This is a bit weird since clients like VSCode aren't actually allowed to send exit-notification until after they've received a response to shutdown-request, and they're not allowed to EOF until after they've sent exit-notification. So it's a scenario that arises in "blind test harnesses" only. (blind meaning a test harness that fires the full volley of messages, rather than waiting for one before sending the next).

Anyway, for this scenario to work, the main process should receive the hover and respond to it, then receive the shutdown-request and respond to it, then receive the exit-notification and shut down -- and never even try to read the EOF that follows.

The problem was that the daemon was getting the EOF from stdin, and dropping whatever outstanding messages were in the daemon's queue, and simply sending the EOF over the pipe to the main process, and then closing the pipe.

The solution is that when the daemon gets an EOF from stdin, then it should discharge the remaining messages in its daemon-queue over the pipe before acting on the EOF. That's what this diff does.

THIS DIFF IS A HACK! I wrote it out just as a proof of the minimal change needed to make the thing work. The next diff in the stack rewrites it cleanly. (but it's a more involved change). Please read this diff solely for sake of understanding the exact bug scenario.

Reviewed By: Wilfred

Differential Revision: D34847716

fbshipit-source-id: 980ab32708d0957f15275663f3c5f909d3e224eb